+++ src/main/java/com/cheftory/api/account/auth/entity/LoggedIn.java
