+++ src/main/java/com/cheftory/api/account/auth/exception/AuthException.java
