+++ src/main/java/com/cheftory/api/account/auth/jwt/TokenProvider.java
@@ -1,8 +1,8 @@
 package com.cheftory.api.auth.jwt;
 
-import com.cheftory.api.auth.config.JwtProperties;
 import com.cheftory.api.auth.exception.AuthErrorCode;
 import com.cheftory.api.auth.exception.AuthException;
+import com.cheftory.api.auth.jwt.property.JwtProperties;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
