+++ src/main/java/com/cheftory/api/account/auth/model/AuthToken.java
